feat(frontend): get map POIs from API instead of hard coded list
POIs ("plekken") are now fetched from the API, falling back on the hard
coded list if the request failed. One problem is that the API's URL
being fetched depends on the hosting solution. Keeping it at
`http://localhost:4000` is not an option. Perhaps if we finish the
docker setup, we could use the configured hostname/port of the container
instead. That way it should work when developing locally using the
docker-compose configuration and when deployed to the server.

Also, facilities such as wash- and bathrooms really aren't
accommodations at all. I had not accounted for those in the database
(nor has anyone pointed out the flaw to me), so the backend will have to
change to better handle those. For now I've added them to accommodations
table with a price of 0, but it really should be fixed.

+++ api/migrations/000001_create_db.up.sql
@@ -101,7 +101,7 @@ CREATE TABLE IF NOT EXISTS accommodation_features (
 CREATE TABLE IF NOT EXISTS accommodations (
     id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     accommodation_type_id BIGINT UNSIGNED NOT NULL,
-    accommodation_number TINYINT UNSIGNED NOT NULL,
+    accommodation_number SMALLINT UNSIGNED NOT NULL,
     coordinate GEOMETRY NOT NULL,
     created_at DATETIME DEFAULT now(),
     updated_at DATETIME ON UPDATE now(),
